requirements.txt
torch
torchvision
tqdm
numpy
pandas
openai
open_clip
Pillow